tulong-server_/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
numpy==1.24.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
Pillow==9.5.0
PyJWT==2.7.0

